+++ src/fr/maxlego08/menu/MenuItemStack.java
@@ -1,9 +1,12 @@
 package fr.maxlego08.menu;
 
 import fr.maxlego08.menu.api.InventoryManager;
+import fr.maxlego08.menu.api.attribute.Attribute;
 import fr.maxlego08.menu.api.attribute.IAttribute;
 import fr.maxlego08.menu.api.loader.MaterialLoader;
+import fr.maxlego08.menu.api.utils.MapConfiguration;
 import fr.maxlego08.menu.api.utils.Placeholders;
+import fr.maxlego08.menu.exceptions.ItemEnchantException;
 import fr.maxlego08.menu.save.Config;
 import fr.maxlego08.menu.zcore.logger.Logger;
 import fr.maxlego08.menu.zcore.utils.Banner;
@@ -16,8 +19,12 @@ import fr.maxlego08.menu.zcore.utils.meta.Meta;
 import fr.maxlego08.menu.zcore.utils.nms.NMSUtils;
 import fr.maxlego08.menu.zcore.utils.nms.NmsVersion;
 import org.bukkit.Bukkit;
+import org.bukkit.Color;
+import org.bukkit.DyeColor;
 import org.bukkit.Material;
 import org.bukkit.OfflinePlayer;
+import org.bukkit.block.banner.Pattern;
+import org.bukkit.block.banner.PatternType;
 import org.bukkit.enchantments.Enchantment;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemFlag;
@@ -25,14 +32,17 @@ import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.EnchantmentStorageMeta;
 import org.bukkit.inventory.meta.ItemMeta;
 import org.bukkit.inventory.meta.SkullMeta;
+import org.bukkit.potion.PotionType;
 
 import javax.annotation.Nullable;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 public class MenuItemStack extends ZUtils {
 
@@ -191,8 +201,11 @@ public class MenuItemStack extends ZUtils {
             itemStack = new ItemStack(material, amount, (byte) this.data);
         }
 
-        if (this.url != null) {
-            itemStack = this.createSkull(this.papi(this.url, player));
+        if (this.url != null && !url.equalsIgnoreCase("null")) {
+            String urlResult = this.papi(this.url, player);
+            if (urlResult != null) {
+                itemStack = this.createSkull(urlResult);
+            }
         }
 
         if (this.potion != null) {
@@ -573,4 +586,157 @@ public class MenuItemStack extends ZUtils {
         this.needPlaceholderAPI = needPlaceholderAPI;
     }
 
+    public void setTypeMapAccessor(MapConfiguration configuration) {
+
+        setData(configuration.getInt("data", 0));
+        setDurability(configuration.getInt("durability", 0));
+        setAmount(configuration.getString("amount", "1"));
+        setMaterial(configuration.getString("material", null));
+        setTargetPlayer(configuration.getString("target", null));
+        setUrl(configuration.getString("url", null));
+
+        Color potionColor = getColor(configuration, "color", null);
+
+        try {
+            Material material = Material.valueOf(configuration.getString("material", "").toUpperCase());
+            String materialName = material.toString();
+            if (materialName.startsWith("LEATHER_")) {
+                Color armorColor = getColor(configuration, "color", Color.fromRGB(160, 101, 64));
+                String type = materialName.replace("LEATHER_", "");
+                setLeatherArmor(new LeatherArmor(LeatherArmor.ArmorType.valueOf(type), armorColor));
+            }
+        } catch (Exception ignored) {
+        }
+
+        if (configuration.contains("potion")) {
+            PotionType type = PotionType.valueOf(configuration.getString("potion", "REGEN").toUpperCase());
+            int level = configuration.getInt("level", 1);
+            boolean splash = configuration.getBoolean("splash", false);
+            boolean extended = configuration.getBoolean("extended", false);
+
+            Potion potion = new Potion(type, level, splash, extended);
+            potion.setColor(potionColor);
+            setPotion(potion);
+        }
+
+        if (configuration.contains("banner")) {
+
+            DyeColor dyeColor = DyeColor.valueOf(configuration.getString("banner", "WHITE").toUpperCase());
+            List<String> stringPattern = configuration.getStringList("patterns");
+            List<Pattern> patterns = new ArrayList<>();
+            for (String string : stringPattern) {
+                String[] split = string.split(":");
+                if (split.length != 2) continue;
+                patterns.add(new Pattern(DyeColor.valueOf(split[0]), PatternType.valueOf(split[1])));
+            }
+            setBanner(new Banner(dyeColor, patterns));
+
+        }
+
+        /*if (configuration.contains("firework")) {
+
+            ConfigurationSection section = configuration.getConfigurationSection("firework");
+            if (section != null) {
+                boolean isStar = section.getBoolean("star", false);
+                FireworkEffect.Builder builder = FireworkEffect.builder();
+                builder.flicker(section.getBoolean("flicker"));
+                builder.trail(section.getBoolean("trail"));
+                builder.with(FireworkEffect.Type.valueOf(section.getString("type", "BALL")));
+                builder.withColor(getColors(section, "colors"));
+                builder.withFade(getColors(section, "fadeColors"));
+                setFirework(new Firework(isStar, builder.build()));
+            }
+        }*/
+
+        List<String> lore = configuration.getStringList("lore");
+        if (lore.isEmpty()) {
+            Object object = configuration.getObject("lore", null);
+            if (object instanceof String) {
+                String loreString = (String) object;
+                lore = Arrays.asList(loreString.split("\n"));
+            }
+        }
+        setLore(lore);
+        setDisplayName(configuration.getString("name", null));
+        setGlowing(configuration.getBoolean("glow"));
+        setModelID(configuration.getString("modelID", configuration.getString("modelId", configuration.getString("customModelId", configuration.getString("customModelData", "0")))));
+
+        List<String> enchants = configuration.getStringList("enchants");
+        Map<Enchantment, Integer> enchantments = new HashMap<>();
+
+        for (String enchantString : enchants) {
+
+            try {
+
+                String[] splitEnchant = enchantString.split(",");
+
+                if (splitEnchant.length == 1)
+                    throw new ItemEnchantException("an error occurred while loading the enchantment " + enchantString);
+
+                int level;
+                String enchant = splitEnchant[0];
+                try {
+                    level = Integer.parseInt(splitEnchant[1]);
+                } catch (NumberFormatException e) {
+                    throw new ItemEnchantException("an error occurred while loading the enchantment " + enchantString);
+                }
+
+                Enchantment enchantment = Enchantment.getByName(enchant);
+                if (enchantment == null) {
+                    throw new ItemEnchantException("an error occurred while loading the enchantment " + enchantString);
+                }
+
+                enchantments.put(enchantment, level);
+
+            } catch (ItemEnchantException e) {
+                e.printStackTrace();
+            }
+        }
+
+        List<ItemFlag> flags = configuration.getStringList("flags").stream().map(this::getFlag).collect(Collectors.toList());
+
+        List<IAttribute> attributeModifiers = new ArrayList<>();
+
+        if (configuration.contains("attributes")) {
+            List<Map<String, Object>> attributesFromConfig = (List<Map<String, Object>>) configuration.getList("attributes");
+            if (attributesFromConfig != null) {
+                for (Map<String, Object> attributeMap : attributesFromConfig) {
+                    attributeModifiers.add(Attribute.deserialize(attributeMap));
+                }
+            }
+        }
+
+        setEnchantments(enchantments);
+        setFlags(flags);
+        setAttributes(attributeModifiers);
+
+    }
+
+    private Color getColor(MapConfiguration configuration, String key, Color def) {
+        String[] split = configuration.getString(key, "").split(",");
+
+        if (split.length == 3) {
+            return Color.fromRGB(Integer.parseInt(split[0]), Integer.parseInt(split[1]), Integer.parseInt(split[2]));
+        } else if (split.length == 4) {
+            return Color.fromARGB(Integer.parseInt(split[0]), Integer.parseInt(split[1]), Integer.parseInt(split[2]), Integer.parseInt(split[3]));
+        }
+
+        return def;
+    }
+
+    private List<Color> getColors(MapConfiguration section, String key) {
+        List<Color> colors = new ArrayList<>();
+
+        for (String color : section.getStringList(key)) {
+            String[] split = color.split(",");
+
+            if (split.length == 3) {
+                colors.add(Color.fromRGB(Integer.parseInt(split[0]), Integer.parseInt(split[1]), Integer.parseInt(split[2])));
+            } else if (split.length == 4) {
+                colors.add(Color.fromARGB(Integer.parseInt(split[0]), Integer.parseInt(split[1]), Integer.parseInt(split[2]), Integer.parseInt(split[3])));
+            }
+        }
+
+        return colors;
+    }
 }